% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Annoy-class.R
\name{AnnoyParam}
\alias{AnnoyParam}
\title{The AnnoyParam class}
\usage{
AnnoyParam(
  ntrees = 50,
  directory = tempdir(),
  search.mult = ntrees,
  distance = "Euclidean"
)
}
\arguments{
\item{ntrees}{Integer scalar, number of trees to use for index generation.}

\item{directory}{String, the directory in which to save the index.}

\item{search.mult}{Numeric scalar, multiplier for the number of points to search.}

\item{distance}{String, the distance metric to use.
Defaults to \code{"Euclidean"}.}
}
\value{
An AnnoyParam object.
}
\description{
A class to hold parameters for the Annoy algorithm for approximate nearest neighbor identification.
}
\details{
The AnnoyParam class holds any parameters associated with running the Annoy algorithm.
This generally relates to building of the index - see \code{\link{buildAnnoy}} for details.
}
\examples{
out <- AnnoyParam()
AnnoyParam_ntrees(out) 
AnnoyParam_directory(out) 

}
\seealso{
\code{\link{buildAnnoy}}.
}
\author{
Aaron Lun
}
