% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiocNeighbors-package.R
\docType{package}
\name{BiocNeighbors-package}
\alias{BiocNeighbors}
\alias{BiocNeighbors-package}
\title{BiocNeighbors: Nearest Neighbor Detection for Bioconductor Packages}
\description{
Implements exact and approximate methods for nearest neighbor detection, in a framework that allows them to be easily switched within Bioconductor packages or workflows. Exact searches can be performed using the k-means for k-nearest neighbors algorithm, vantage point trees, or an exhaustive search. Approximate searches can be performed using the Annoy, HNSW or FLANN libraries, nearest-neighbor descent, or random projection forests. Each search can be performed with a variety of different distance metrics, parallelization, and variable numbers of neighbors. Range-based searches (to find all neighbors within a certain distance) are also supported.
}
\author{
\strong{Maintainer}: Aaron Lun \email{infinite.monkeys.with.keyboards@gmail.com} [copyright holder]

}
