% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiocNeighborParam-class.R
\name{bndistance,BiocNeighborParam-method}
\alias{bndistance,BiocNeighborParam-method}
\title{Return a string specifying the distance metric to be used for searching, usually \code{"Euclidean"} or \code{"Manhattan"}.}
\usage{
\S4method{bndistance}{BiocNeighborParam}(x)
}
\description{
Return a string specifying the distance metric to be used for searching, usually \code{"Euclidean"} or \code{"Manhattan"}.
}
