% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert}
\alias{convert}
\alias{convertToSparseMatrix}
\alias{convertToSelfHits}
\title{Converts \code{findKNN} output to alternative representations}
\usage{
convertToSparseMatrix(knn, repr = "C")

convertToSelfHits(knn)
}
\arguments{
\item{knn}{Named list containing the output from \code{\link{findKNN}}.
This list should have the \code{index} and \code{distance} matrices.}

\item{repr}{String specifing the type of sparse matrix, see \code{?\link[Matrix]{sparseMatrix}} for more details.}
}
\value{
For \code{convertToSparseMatrix}, a floating-point sparse matrix from the \pkg{Matrix} package.
Each row corresponds to an observation and the columns with non-zero entries represent the neighbors.
The value of each non-zero entry contains the distance to that neighbor.

For \code{convertToSelfHits}, a \link[S4Vectors]{SelfHits} object where \code{from} contains the observations and \code{to} contains the identities of the neighbors.
The \code{distance} metadata field contains the distance from each observation to each of its neighbors.
}
\description{
Convert the output of \code{\link{findKNN}} to alternative representations for more convenient downstream use.
}
\examples{
X <- matrix(rnorm(10000), ncol=20)
out <- findKNN(X, k=5)

sm <- convertToSparseMatrix(out)
str(sm)

sh <- convertToSelfHits(out)
sh

}
\seealso{
The \code{colPairs} slot of the SingleCellExperiment class, from the \pkg{SingleCellExperiment} package.
}
