% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiocNeighborParam-class.R
\name{show,BiocNeighborParam-method}
\alias{show,BiocNeighborParam-method}
\title{Display the class of a BiocNeighborParam \code{object}.}
\usage{
\S4method{show}{BiocNeighborParam}(object)
}
\description{
Display the class of a BiocNeighborParam \code{object}.
}
