% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Vptree-class.R
\name{VptreeParam}
\alias{VptreeParam}
\title{The VptreeParam class}
\usage{
VptreeParam(distance = "Euclidean")
}
\arguments{
\item{distance}{A string specifying the distance metric to use.
Defaults to \code{"Euclidean"}.}
}
\value{
A VptreeParam object.
}
\description{
A class to hold parameters for the VP tree algorithm for exact nearest neighbor identification.
}
\examples{
out <- VptreeParam()

}
\seealso{
\code{\link{buildVptree}}.
}
\author{
Aaron Lun
}
