% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Annoy-class.R
\name{AnnoyIndex}
\alias{AnnoyIndex}
\alias{show,AnnoyIndex-method}
\alias{AnnoyIndex_path}
\alias{AnnoyIndex_search_mult}
\alias{bnorder,AnnoyIndex-method}
\title{The AnnoyIndex class}
\usage{
AnnoyIndex(data, path, search.mult = 50, NAMES = NULL, distance = "Euclidean")

\S4method{show}{AnnoyIndex}(object)

AnnoyIndex_path(x)

AnnoyIndex_search_mult(x)

\S4method{bnorder}{AnnoyIndex}(x)
}
\arguments{
\item{data}{A numeric matrix with data points in columns and dimensions in rows.}

\item{path}{A string specifying the path to the index file.}

\item{search.mult}{Numeric scalar, multiplier for the number of points to search.}

\item{NAMES}{A character vector of sample names or \code{NULL}.}

\item{distance}{A string specifying the distance metric to use.}
}
\value{
An AnnoyIndex object.
}
\description{
A class to hold indexing structures for the Annoy algorithm for approximate nearest neighbor identification.
}
\details{
The AnnoyIndex class holds the indexing structure required to run the Annoy algorithm.
Users should never need to call the constructor explicitly, but should generate instances of AnnoyIndex classes with \code{\link{buildAnnoy}}.
}
\examples{
out <- example(buildAnnoy)
str(AnnoyIndex_path(out))
str(AnnoyIndex_search_mult(out))

}
\seealso{
\code{\link{buildAnnoy}}.
}
\author{
Aaron Lun
}
