% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiocNeighborParam-class.R
\docType{class}
\name{BiocNeighborParam}
\alias{BiocNeighborParam}
\alias{BiocNeighborParam,}
\alias{BiocNeighborParam-class,}
\alias{bndistance,BiocNeighborParam-method,}
\alias{show,BiocNeighborParam-method}
\alias{show,BiocNeighborParam}
\alias{bndistance,BiocNeighborParam-method}
\alias{bndistance,BiocNeighborParam}
\title{The BiocNeighborParam class}
\usage{
\S4method{show}{BiocNeighborParam}(object)

\S4method{bndistance}{BiocNeighborParam}(x)
}
\description{
A virtual class for specifying the type of nearest-neighbor search algorithm and associated parameters.
}
\details{
The BiocNeighborParam class is a virtual base class on which other parameter objects are built.
There are currently 5 concrete subclasses:
\describe{
    \item{}{\code{\link{ExhaustiveParam}}: exact nearest-neighbor search with the exhaustive (i.e. brute-force) algorithm.}
    \item{}{\code{\link{KmknnParam}}: exact nearest-neighbor search with the KMKNN algorithm.}
    \item{}{\code{\link{VptreeParam}}: exact nearest-neighbor search with the VP tree algorithm.}
    \item{}{\code{\link{AnnoyParam}}: approximate nearest-neighbor search with the Annoy algorithm.}
    \item{}{\code{\link{HnswParam}}: approximate nearest-neighbor search with the HNSW algorithm.}
These objects hold parameters specifying how each algorithm should be run on an arbitrary data set.
See the associated documentation pages for more details.
}
}
\seealso{
\code{\link{ExhaustiveParam}},
\code{\link{KmknnParam}},
\code{\link{VptreeParam}},
\code{\link{AnnoyParam}},
and \code{\link{HnswParam}} for constructors.

\code{\link{buildIndex}}, \code{\link{findKNN}} and \code{\link{queryKNN}} for dispatch.
}
\author{
Aaron Lun
}
