% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/defineBuilder.R
\name{defineBuilder}
\alias{defineBuilder}
\alias{defineBuilder,NULL-method}
\alias{defineBuilder,missing-method}
\title{Define an index builder}
\usage{
defineBuilder(BNPARAM)
}
\arguments{
\item{BNPARAM}{A \linkS4class{BiocNeighborParam} object specifying the type of index to be constructed.
If \code{NULL}, this defaults to a \linkS4class{KmknnParam} object.}
}
\value{
Pointer to a builder instance that can be used to construct a prebuilt index in \code{\link{buildIndex}}.
}
\description{
Define a builder object that can construct indices for nearest-neighbor searching with different algorithms.
}
\details{
Methods for this generic should return a pointer that refers to a \code{BiocNeighbors::Builder} object,
see definition in \code{system.file("include", "BiocNeighbors.h", package="BiocNeighbors")} for details.

Note that the pointer returned by \code{defineBuilder} should \emph{not} be used in \code{\link{findKNN}}, \code{\link{queryKNN}}, etc.
Those methods instead accept the pointer returned by \code{\link{buildIndex}}.
}
\examples{
(out <- defineBuilder())
(out2 <- defineBuilder(AnnoyParam()))

}
\seealso{
\code{\link{defineBuilder,KmknnParam-method}},
\code{\link{defineBuilder,VptreeParam-method}},
\code{\link{defineBuilder,AnnoyParam-method}} 
and \code{\link{defineBuilder,HnswParam-method}} for specific methods.
}
\author{
Aaron Lun
}
