% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExhaustiveParam-class.R
\docType{class}
\name{ExhaustiveParam}
\alias{ExhaustiveParam}
\alias{ExhaustiveParam-class}
\title{The ExhaustiveParam class}
\usage{
ExhaustiveParam(distance = "Euclidean")
}
\arguments{
\item{distance}{A string specifying the distance metric to use.}
}
\value{
An instance of the ExhaustiveParam class.
}
\description{
A class to hold parameters for the exhaustive algorithm for exact nearest neighbor identification.
}
\examples{
(out <- ExhaustiveParam())

}
\seealso{
\code{\link{buildExhaustive}}, for the index construction.

\code{\link{findExhaustive}} and related functions, for the actual search. 

\linkS4class{BiocNeighborParam}, for the parent class and its available methods.
}
\author{
Allison Vuong
}
