% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hnsw-class.R
\name{HnswParam}
\alias{HnswParam}
\alias{HnswParam_nlinks}
\alias{HnswParam_ef_construction}
\alias{HnswParam_directory}
\alias{HnswParam_ef_search}
\alias{show,HnswParam-method}
\title{The HnswParam class}
\usage{
HnswParam(
  nlinks = 16,
  ef.construction = 200,
  directory = tempdir(),
  ef.search = 10,
  distance = "Euclidean"
)

HnswParam_nlinks(x)

HnswParam_ef_construction(x)

HnswParam_directory(x)

HnswParam_ef_search(x)

\S4method{show}{HnswParam}(object)
}
\arguments{
\item{nlinks}{Integer scalar, number of bi-directional links per element for index generation.}

\item{ef.construction}{Integer scalar, size of the dynamic list for index generation.}

\item{directory}{String, the directory in which to save the index.}

\item{ef.search}{Integer scalar, size of the dynamic list for neighbor searching.}

\item{distance}{String, the distance metric to use.
Defaults to \code{"Euclidean"}.}
}
\value{
An HnswParam object.
}
\description{
A class to hold parameters for the Hnsw algorithm for approximate nearest neighbor identification.
}
\details{
The HnswParam class holds any parameters associated with running the HNSW algorithm.
This generally relates to building of the index - see \code{\link{buildHnsw}} for details.
}
\examples{
out <- HnswParam()
HnswParam_nlinks(out) 
HnswParam_ef_construction(out) 
HnswParam_directory(out) 

}
\seealso{
\code{\link{buildHnsw}}.
}
\author{
Aaron Lun
}
