% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kmknn-class.R
\name{KmknnIndex}
\alias{KmknnIndex}
\alias{show,KmknnIndex-method}
\alias{KmknnIndex_cluster_centers}
\alias{KmknnIndex_cluster_info}
\alias{bnorder,KmknnIndex-method}
\alias{KmknnIndex-method}
\title{The KmknnIndex class}
\usage{
KmknnIndex(data, centers, info, order, NAMES = NULL, distance = "Euclidean")

\S4method{show}{KmknnIndex}(object)

KmknnIndex_cluster_centers(x)

KmknnIndex_cluster_info(x)

\S4method{bnorder}{KmknnIndex}(x)
}
\arguments{
\item{data}{A numeric matrix with data points in columns and dimensions in rows.}

\item{centers}{A numeric matrix with clusters in columns and dimensions in rows.}

\item{info}{A list of statistics for each cluster.}

\item{order}{An integer vector of length equal to \code{ncol(data)}, specifying the order of observations.}

\item{NAMES}{A character vector of sample names or \code{NULL}.}

\item{distance}{A string specifying the distance metric to use.}
}
\value{
A KmknnIndex object.
}
\description{
A class to hold indexing structures for the KMKNN algorithm for exact nearest neighbor identification.
}
\details{
The KmknnIndex class holds the indexing structure required to run the KMKNN algorithm.
Users should never need to call the constructor explicitly, but should generate instances of KmknnIndex classes with \code{\link{buildKmknn}}.
}
\examples{
out <- KmknnIndex(data, centers, info, order, NAMES=NULL, distance="Euclidean")
str(KmknnIndex_cluster_centers(out))
str(KmknnIndex_cluster_info(out))

}
\seealso{
\code{\link{buildKmknn}}.
}
\author{
Aaron Lun
}
