% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kmknn-class.R
\name{KmknnParam}
\alias{KmknnParam}
\alias{show,KmknnParam-method}
\title{The KmknnParam class}
\usage{
KmknnParam(..., distance = "Euclidean")

\S4method{show}{KmknnParam}(object)
}
\arguments{
\item{...}{Arguments to be passed to \code{\link{kmeans}}.}

\item{distance}{A string specifying the distance metric to use.
Defaults to \code{"Euclidean"}.}
}
\value{
A KmknnParam object.
}
\description{
A class to hold parameters for the KMKNN algorithm for exact nearest neighbor identification.
}
\details{
The KmknnParam class holds any parameters associated with running the KMKNN algorithm.
Currently, this relates to tuning of the k-means step - see \code{\link{buildKmknn}} for details.
}
\examples{
out <- KmknnParam(iter.max=100)
KmknnParam_kmeans_args(out) 

}
\seealso{
\code{\link{buildKmknn}}.
}
\author{
Aaron Lun
}
