% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryKNN.R
\name{queryKNN}
\alias{queryKNN}
\title{Query nearest neighbors}
\usage{
queryKNN(
  X,
  query,
  k,
  get.index = TRUE,
  get.distance = TRUE,
  num.threads = 1,
  subset = NULL,
  transposed = FALSE,
  ...,
  BPPARAM = NULL
)
}
\arguments{
\item{X}{A numeric matrix where rows correspond to data points and columns correspond to variables (i.e., dimensions).
Alternatively, a pointer to a prebuilt index from \code{\link{buildIndex}}.}

\item{query}{A numeric matrix of query points, containing the same number of columns as \code{X}.}

\item{k}{A positive integer scalar specifying the number of nearest neighbors to retrieve.}

\item{get.index}{A logical scalar indicating whether the indices of the nearest neighbors should be recorded.
Setting this to \code{FALSE} improves efficiency if the indices are not of interest.

Alternatively, this may be a string containing \code{"normal"} or \code{"transposed"}.
The former is the same as \code{TRUE}, while the latter returns the index matrix in transposed format.}

\item{get.distance}{A logical scalar indicating whether distances to the nearest neighbors should be recorded.
Setting this to \code{FALSE} improves efficiency if the distances are not of interest.

Alternatively, this may be a string containing \code{"normal"} or \code{"transposed"}.
The former is the same as \code{TRUE}, while the latter returns the distance matrix in transposed format.}

\item{num.threads}{Integer scalar specifying the number of threads to use for the search.}

\item{subset}{A vector indicating the rows of \code{query} for which the nearest neighbors should be identified.}

\item{transposed}{A logical scalar indicating whether \code{X} and \code{query} are transposed, 
in which case both matrices are assumed to contain dimensions in the rows and data points in the columns.}

\item{...}{Further arguments to pass to \code{\link{buildIndex}} when \code{X} is not an external pointer.
This includes \code{BNPARAM} to specify how the index should be constructed.}

\item{BPPARAM}{Deprecated, use \code{num.threads} instead.}
}
\value{
A list is returned containing:
\itemize{
\item \code{index}, if \code{get.index=TRUE}.
This is an integer matrix where each row corresponds to a point (denoted here as \eqn{i}) in \code{query}.
The row for \eqn{i} contains the row indices of \code{X} that are the nearest neighbors to point \eqn{i}, sorted by increasing distance from \eqn{i}.
\item \code{distance}, if \code{get.distance=TRUE}.
This is a numeric matrix where each row corresponds to a point (as above) and contains the sorted distances of the neighbors from \eqn{i}.
}

The number of columns in both matrices is set to \code{min(k, ncol(X))}.
If \code{subset} is not \code{NULL}, each row of the above matrices refers to a point in the subset, in the same order as supplied in \code{subset}.

If \code{get.index="transposed"}, the \code{index} matrix is transposed, i.e., the rows are the neighbors while the columns are the points.
Similarly, if \code{get.distance="transposed"}, the \code{distance} matrix is transposed.
}
\description{
Query a dataset for nearest neighbors of points in another dataset, using a variety of algorithms.
}
\details{
If multiple queries are to be performed to the same \code{X}, it may be beneficial to build the index from \code{X} with \code{\link{buildIndex}}.
The resulting pointer object can be supplied as \code{X} to multiple \code{queryKNN} calls, avoiding the need to repeat index construction in each call.
}
\examples{
Y <- matrix(rnorm(100000), ncol=20)
Z <- matrix(rnorm(20000), ncol=20)
out <- queryKNN(Y, query=Z, k=5)
head(out$index)
head(out$distance)

}
\seealso{
\code{\link{buildIndex}}, to build an index ahead of time.
}
\author{
Aaron Lun
}
