% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Vptree-class.R
\name{VptreeIndex}
\alias{VptreeIndex}
\alias{VptreeIndex_nodes}
\alias{bnorder,VptreeIndex-method}
\alias{VptreeIndex-method}
\title{The VptreeIndex class}
\usage{
VptreeIndex(data, nodes, order, NAMES = NULL, distance = "Euclidean")

VptreeIndex_nodes(x)

\S4method{bnorder}{VptreeIndex}(x)
}
\arguments{
\item{data}{A numeric matrix with data points in columns and dimensions in rows.}

\item{nodes}{A list of vectors specifying the structure of the VP tree.}

\item{order}{An integer vector of length equal to \code{ncol(data)}, specifying the order of observations.}

\item{NAMES}{A character vector of sample names or \code{NULL}.}

\item{distance}{A string specifying the distance metric to use.}
}
\value{
A VptreeIndex object.
}
\description{
A class to hold the vantage point tree for exact nearest neighbor identification.
}
\details{
The VptreeIndex class holds the indexing structure required to run the Vptree algorithm.
Users should never need to call the constructor explicitly, but should generate instances of VptreeIndex classes with \code{\link{buildVptree}}.
}
\examples{
out <- example(buildVptree)
str(VptreeIndex_nodes(out))

}
\seealso{
\code{\link{buildVptree}}.
}
\author{
Aaron Lun
}
