% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VptreeParam-class.R
\name{VptreeParam}
\alias{VptreeParam}
\alias{VptreeParam-class}
\title{The VptreeParam class}
\usage{
VptreeParam(distance = "Euclidean")
}
\arguments{
\item{distance}{A string specifying the distance metric to use.}
}
\value{
An instance of the VptreeParam class.
}
\description{
A class to hold parameters for the VP tree algorithm for exact nearest neighbor identification.
}
\examples{
(out <- VptreeParam())

}
\seealso{
\code{\link{buildVptree}}, for the index construction.

\code{\link{findVptree}} and related functions, for the actual search. 

\linkS4class{BiocNeighborParam}, for the parent class and its available methods.
}
\author{
Aaron Lun
}
